# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "GIN CRUD",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "assusa456u.local:8802",
    "basePath": "/",
    "paths": {
...
        "/v1/m_user": {
            "get": {
                "description": "Get Page MUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mUser"
                ],
                "summary": "MUserPage",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "default": "0",
                        "description": "page",
                        "name": "_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "5",
                        "description": "size",
                        "name": "_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "_sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter",
                        "name": "_filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "global filter",
                        "name": "_q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create MUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mUser"
                ],
                "summary": "MUserCreate",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "description": "Add MUser",
                        "name": "mUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/m_user/delete/{id}": {
            "put": {
                "description": "Soft Delete MUser by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mUser"
                ],
                "summary": "MUserSoftDelete",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "MUser id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/m_user/header": {
            "get": {
                "description": "Get MUser header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mUser"
                ],
                "summary": "MUserHeader",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/m_user/{id}": {
            "get": {
                "description": "Get MUser by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mUser"
                ],
                "summary": "MUserIndex",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "MUser id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Update MUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mUser"
                ],
                "summary": "MUserUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "description": "Update MUser",
                        "name": "mUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MUser"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "MUser id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete MUser by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mUser"
                ],
                "summary": "MUserDelete",
                "parameters": [
                    {
                        "type": "string",
                        "default": "gzip",
                        "description": "gzip",
                        "name": "Accept-Encoding",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "MUser id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.MBiodata": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "createdBy": {
                    "type": "integer"
                },
                "createdOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "deletedBy": {
                    "type": "integer"
                },
                "deletedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "fullname": {
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "imagePath": {
                    "type": "string",
                    "maxLength": 255
                },
                "isDelete": {
                    "type": "boolean"
                },
                "mobilePhone": {
                    "type": "string",
                    "maxLength": 15
                },
                "modifiedBy": {
                    "type": "integer"
                },
                "modifiedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                }
            }
        },
...
        "model.MRole": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 20
                },
                "createdBy": {
                    "type": "integer"
                },
                "createdOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "deletedBy": {
                    "type": "integer"
                },
                "deletedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "id": {
                    "type": "integer"
                },
                "isDelete": {
                    "type": "boolean"
                },
                "level": {
                    "type": "integer"
                },
                "modifiedBy": {
                    "type": "integer"
                },
                "modifiedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
        "model.MUser": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "biodataId": {
                    "type": "integer"
                },
                "createdBy": {
                    "type": "integer"
                },
                "createdOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "deletedBy": {
                    "type": "integer"
                },
                "deletedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "email": {
                    "type": "string",
                    "maxLength": 100
                },
                "id": {
                    "type": "integer"
                },
                "isDelete": {
                    "type": "boolean"
                },
                "isLocked": {
                    "type": "boolean"
                },
                "lastLogin": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "loginAttempt": {
                    "type": "integer"
                },
                "mbiodata": {
                    "$ref": "#/definitions/model.MBiodata"
                },
                "modifiedBy": {
                    "type": "integer"
                },
                "modifiedOn": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                },
                "mrole": {
                    "$ref": "#/definitions/model.MRole"
                },
                "password": {
                    "type": "string",
                    "maxLength": 255
                },
                "roleId": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "path": {
                    "type": "string",
                    "example": "/v1/m_biodata"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-02-16 10:33:10"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}